benchmarks/code/requirements.txt
attrdict==2.0.1
gensim==3.8.0
matplotlib==3.1.1
# nltk==3.6.6
# numpy==1.22.0
pandas==0.25.1
regex==2019.8.19
# scipy==1.3.1
sympy==1.4
tensorboard==1.12.2
tensorboardX==1.2
# tokenizers==0.8.1rc2
# torch==1.3.0
torchtext==0.4.0
# torchvision==0.2.1
# tqdm==4.36.1
# transformers==3.3.1
